+++ lava_dispatcher/actions/deploy.py
@@ -21,6 +21,63 @@ from lava_dispatcher.actions import BaseAction
 
 
 class cmd_deploy_linaro_image(BaseAction):
-    def run(self, hwpack=None, rootfs=None, image=None, kernel_matrix=None, use_cache=True, rootfstype='ext3'):
+
+    # This is how the schema for parameters should look, but there are bugs in
+    # json_schema_validation that means it doesn't work (see
+    # https://github.com/zyga/json-schema-validator/pull/6).
+
+    ## parameters_schema = {
+    ##     'type': [
+    ##         {
+    ##             'type': 'object',
+    ##             'properties': {
+    ##                 'image': {'type': 'string'},
+    ##                 },
+    ##             'additionalProperties': False,
+    ##             },
+    ##         {
+    ##             'type': 'object',
+    ##             'properties': {
+    ##                 'hwpack': {'type': 'string'},
+    ##                 'rootfs': {'type': 'string'},
+    ##                 'kernel_matrix': {'type': 'string', 'optional': True},
+    ##                 'use_cache': {'type': 'bool', 'optional': True, 'default': True},
+    ##                 'rootfstype': {'type': 'string', 'optional': True, 'default': 'ext3'},
+    ##                 },
+    ##             'additionalProperties': False,
+    ##             },
+    ##         ],
+    ##     }
+
+    parameters_schema = {
+        'type': 'object',
+        'properties': {
+            'hwpack': {'type': 'string', 'optional': True},
+            'rootfs': {'type': 'string', 'optional': True},
+            'image': {'type': 'string', 'optional': True},
+            'kernel_matrix': {'type': 'string', 'optional': True},
+            'use_cache': {'type': 'bool', 'optional': True},
+            'rootfstype': {'type': 'string', 'optional': True},
+            },
+        'additionalProperties': False,
+        }
+
+    @classmethod
+    def validate_parameters(cls, parameters):
+        super(cmd_deploy_linaro_image, cls).validate_parameters(parameters)
+        if 'hwpack' in parameters:
+            if 'rootfs' not in parameters:
+                raise ValueError('must specify rootfs when specifying hwpack')
+            if 'image' in parameters:
+                raise ValueError('cannot specify image and hwpack')
+        elif 'image' not in parameters:
+            raise ValueError('must specify image if not specifying a hwpack')
+        elif 'kernel_matrix' in parameters:
+            raise ValueError('cannot specify kernel_matrix with an image')
+
+    def run(self, hwpack=None, rootfs=None, image=None, kernel_matrix=None,
+            use_cache=True, rootfstype='ext3'):
         self.client.deploy_linaro(
-            hwpack=hwpack, rootfs=rootfs, image=image, kernel_matrix=kernel_matrix, use_cache=use_cache, rootfstype=rootfstype)
+            hwpack=hwpack, rootfs=rootfs, image=image,
+            kernel_matrix=kernel_matrix, use_cache=use_cache,
+            rootfstype=rootfstype)